Move Telescope Controller to core
